requirements.txt
# Kaggle Traffic Management Training Requirements
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
Pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.4.0
requests>=2.28.0
PyYAML>=6.0
